move network/wireguard to netbase/wireguard
