+++ pkg/network/networker.go
@@ -22,6 +22,7 @@ import (
 	"github.com/threefoldtech/zosbase/pkg/environment"
 	"github.com/threefoldtech/zosbase/pkg/gridtypes"
 	"github.com/threefoldtech/zosbase/pkg/gridtypes/zos"
+	"github.com/threefoldtech/zosbase/pkg/netbase/wireguard"
 	"github.com/threefoldtech/zosbase/pkg/network/bootstrap"
 	"github.com/threefoldtech/zosbase/pkg/network/iperf"
 	"github.com/threefoldtech/zosbase/pkg/network/mycelium"
@@ -29,7 +30,6 @@ import (
 	"github.com/threefoldtech/zosbase/pkg/network/options"
 	"github.com/threefoldtech/zosbase/pkg/network/public"
 	"github.com/threefoldtech/zosbase/pkg/network/tuntap"
-	"github.com/threefoldtech/zosbase/pkg/network/wireguard"
 	"github.com/threefoldtech/zosbase/pkg/network/yggdrasil"
 	"github.com/threefoldtech/zosbase/pkg/stubs"
 	"github.com/threefoldtech/zosbase/pkg/zinit"
