+++ pkg/network/nr/net_resource.go
@@ -29,9 +29,9 @@ import (
 	"github.com/rs/zerolog/log"
 	"github.com/threefoldtech/zosbase/pkg"
 	"github.com/threefoldtech/zosbase/pkg/netbase/nft"
+	"github.com/threefoldtech/zosbase/pkg/netbase/wireguard"
 	"github.com/threefoldtech/zosbase/pkg/network/bridge"
 	"github.com/threefoldtech/zosbase/pkg/network/namespace"
-	"github.com/threefoldtech/zosbase/pkg/network/wireguard"
 	"github.com/vishvananda/netlink"
 
 	baseifaceutil "github.com/threefoldtech/zosbase/pkg/netbase/ifaceutil"
@@ -91,7 +91,6 @@ func (m *MyceliumInspection) Gateway() (subnet net.IPNet, gw net.IPNet, err erro
 }
 
 func (m *MyceliumInspection) IP(seed zos.Bytes) (ip net.IPNet, gw net.IPNet, err error) {
-
 	if slices.ContainsFunc(invalidMyceliumSeeds, func(b []byte) bool {
 		return slices.Equal(seed, b)
 	}) {
@@ -132,7 +131,7 @@ type NetResource struct {
 // is stored.
 func New(nr pkg.Network, keyDir string) *NetResource {
 	return &NetResource{
-		//fix here
+		// fix here
 		id:             nr.NetID,
 		resource:       nr,
 		networkIPRange: nr.NetworkIPRange.IPNet,
@@ -328,7 +327,6 @@ func (nr *NetResource) SetMycelium() (err error) {
 	err = zinit.AddService(name, zinit.InitService{
 		Exec: strings.Join(args, " "),
 	})
-
 	if err != nil {
 		return errors.Wrap(err, "failed to add mycelium service for nr")
 	}
@@ -419,7 +417,6 @@ func (nr *NetResource) ensureMyceliumNetwork(keyFile string) error {
 
 		return nil
 	})
-
 }
 
 func (nr *NetResource) inspectMycelium(keyFile string) (inspection MyceliumInspection, err error) {
@@ -496,7 +493,6 @@ func (nr *NetResource) ConfigureWG(privateKey string) error {
 	}
 
 	handler := func(_ ns.NetNS) error {
-
 		wgName, err := nr.WGName()
 		if err != nil {
 			return err
@@ -635,7 +631,6 @@ func (nr *NetResource) Delete() error {
 }
 
 func (nr *NetResource) wgPeers() ([]*wireguard.Peer, error) {
-
 	wgPeers := make([]*wireguard.Peer, 0, len(nr.resource.Peers)+1)
 
 	for _, peer := range nr.resource.Peers {
@@ -718,7 +713,7 @@ func (nr *NetResource) attachToNRBridge() error {
 		}
 	}
 
-	var handler = func(_ ns.NetNS) error {
+	handler := func(_ ns.NetNS) error {
 		link, err := netlink.LinkByName(nrIfaceName)
 		if err != nil {
 			return err
@@ -880,5 +875,4 @@ func AppendFunc[A any, B any, S []A, D []B](d D, s S, f func(A) B) D {
 	}
 
 	return d
-
 }
